change in more_view
