+++ lib/features/more/widget/about_applets_tile.dart
@@ -6,17 +6,24 @@ import 'package:notredame/features/more/more_viewmodel.dart';
 import 'package:notredame/utils/locator.dart';
 import 'package:notredame/features/app/analytics/analytics_service.dart';
 
-class ReportBugListTile extends ViewModelWidget<MoreViewModel> {
+class AboutAppletsTile extends ViewModelWidget<MoreViewModel> {
   final AnalyticsService _analyticsService = locator<AnalyticsService>();
 
   @override
   Widget build(BuildContext context, MoreViewModel model) {
     return ListTile(
-      title: Text(AppIntl.of(context)!.more_report_bug),
-      leading: const Icon(Icons.bug_report),
+      title: Text(AppIntl.of(context)!.more_about_applets_title),
+      leading: Hero(
+        tag: 'about',
+        child: Image.asset(
+          "assets/images/favicon_applets.png",
+          height: 24,
+          width: 24,
+        ),
+      ),
       onTap: () {
-        _analyticsService.logEvent("MoreView", "Report a bug clicked");
-        model.navigationService.pushNamed(RouterPaths.feedback);
+        _analyticsService.logEvent("MoreView", "About App|ETS clicked");
+        model.navigationService.pushNamed(RouterPaths.about);
       },
     );
   }
