--- lib/features/more/widget/in_app_review_list_tile
@@ -1,22 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:stacked/stacked.dart';
-import 'package:notredame/features/more/more_viewmodel.dart';
-import 'package:notredame/utils/locator.dart';
-import 'package:notredame/features/app/analytics/analytics_service.dart';
-
-class InAppReviewListTile extends ViewModelWidget<MoreViewModel> {
-  final AnalyticsService _analyticsService = locator<AnalyticsService>();
-
-  @override
-  Widget build(BuildContext context, MoreViewModel model) {
-    return ListTile(
-      title: Text(AppIntl.of(context)!.in_app_review_title),
-      leading: Icon(Icons.rate_review),
-      onTap: () {
-        _analyticsService.logEvent("MoreView", "Rate us clicked");
-        MoreViewModel.launchInAppReview();
-      },
-    );
-  }
-}
